dev-requirements.txt
# Keep in-sync w/ cluster/server/requirements.txt, cluster/register/requirements.txt, and client/requirements.txt

# From server/requirements.txt
azure-storage==0.36.0
Flask==1.0.2
Flask-JWT-Extended==3.17.0
gevent==1.4.0
google-cloud-storage==1.13.0
gunicorn==19.9.0
kubernetes==9.0.0
numpy==1.16.1
paramiko==2.4.2
Pillow==5.4.1
psycopg2==2.7.6.1
redis==3.2.0
rq==0.13.0

# From registration-pipelines
# Copied from https://github.com/tensorflow/datasets/blob/master/setup.py
matplotlib # cats_vs_dogs
Pillow # colorectal_histology
numpy==1.16.1
scipy==1.3.0 # quickdraw, etc.
tensorflow-datasets==1.0.2
tensorflow==1.13.1

# Dev dependencies
pytest==4.2.1
pytest-cov==2.7.1
testing.postgresql==1.3.0
flake8==3.7.7